added login required to views

+++ packages/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from .models import Packages
 from .forms import PackagesForm
 
@@ -28,8 +29,15 @@ def package_detail(request, package_id):
     return render(request, 'package_detail.html', context)
 
 
+@login_required
 def add_package(request):
     """ Add a product to the store """
+
+    if not request.user.is_superuser:
+        messages.success(
+            request, 'Sorry, only store owners can access this page.')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = PackagesForm(request.POST, request.FILES)
         if form.is_valid():
@@ -52,8 +60,15 @@ def add_package(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_package(request, package_id):
     """ Edit a package """
+
+    if not request.user.is_superuser:
+        messages.success(
+            request, 'Sorry, only store owners can access this page.')
+        return redirect(reverse('home'))
+
     package = get_object_or_404(Packages, pk=package_id)
     if request.method == 'POST':
         form = PackagesForm(request.POST, request.FILES, instance=package)
@@ -78,8 +93,15 @@ def edit_package(request, package_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_package(request, package_id):
     """ Delete a package """
+
+    if not request.user.is_superuser:
+        messages.success(
+            request, 'Sorry, only store owners can access this page.')
+        return redirect(reverse('home'))
+
     package = get_object_or_404(Packages, pk=package_id)
     package.delete()
     messages.success(request, 'Package deleted!')
